Auto-reconnect when tracker connection is lost
This happens whenever my computer goes to sleep while ydnc-time is
running; because the connection is lost while my program is suspended,
it never receives the `DeviceDisconnected` event. So, whenever the
application _is_ running (in this case, when it has been resumed after
the computer wakes from sleep), if we do not receive a side change
notification from the tracker in 5 seconds, do a quick check on
`.is_connected()` to see if the connection is still live, reconnecting
it if it is not.

If the reconnection fails, fallback to the loop/retry behavior in
spawn_sub_task, which will drop all the subscription handles and attempt
to recreate them after 5s.